feat: add support for Java 17 and Spring 6, restructure project
- Add support for Java 17 and Spring Framework 6
- Restructure project into a multi-module Maven project
- Move smoke tests into their own module, `smoke-tests` for testing compatibility with different Java versions
- Update README with compatibility notes
