--- src/test/resources/logback.xml
@@ -1,17 +1,0 @@
-<configuration>
-
-	<appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-		<encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
-			<pattern>%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n</pattern>
-		</encoder>
-	</appender>
-
-	<logger name="io.florianlopes" level="info" additivity="false">
-		<appender-ref ref="CONSOLE"/>
-	</logger>
-
-	<root level="error">
-		<appender-ref ref="CONSOLE"/>
-	</root>
-
-</configuration>
